AFL utility functions

AFL utility functions

AFL utility functions

Details

Normally you don't need these functions. But they are flexible in constructing high-order AFL expressions.

See also

Methods

Public methods

AFLUtils$e_to_afl_filter()

AFLUtils$args_to_expressions()

AFLUtils$e()

AFLUtils$e_merge()

Method e_to_afl_filter()

Convert filter expression(s) to AFL filter

Usage

AFLUtils$e_to_afl_filter(e, regex_func = "rsub", ignore_case = TRUE)

Arguments

e

An R expression vector of length 1 or more

Returns

An AFL filter string

Method args_to_expressions()

Convert API ... args to an R expression vector

Some API functions include ... arg to represent arbitrary search criteria. This provides flexibility and simplifies API function signatures, but only supports limited advanced search, e.g. xxx_contains, xxx_range, xxx_not.

Eg. name_contains = 'str' => name %contains% 'str' Eg. value_range = c(1, 9) => c(value >= 1, value <= 9)

Usage

AFLUtils$args_to_expressions(..., .dots = NULL)

Arguments

...

API ellipsis arg

.dots

Explicitly provide a parameter list. If not NULL, the ellipsis params are ignored

Returns

R expression vector

Method e()

Create a list of R expressions

The ... ellipsis arg can include arbitrary expressions, where all names are preserved in their literal forms, except for those prefixed with !! (double exclamation marks) which will be evaluated to their actual values in the calling environment.

Besides common comparison operators including ==, >, <, >=, <=, !=, there are a few special operators supported to ease AFL generation:

%in% semantically similar to R. a %in% !!c(1,2,3) will be translated to (a == 1 or a == 2 or a == 3)

%like% for string regex matching.

Usage

AFLUtils$e(...)

Arguments

...

The ellipsis arg can have multiple items as expressions, but NO named items as in a named list.

Returns

A list of R expressions

Method e_merge()

Merge multiple R expressions into one

Merge an ExprsList into a single Expression so that it can be used as a FilterExpr

Usage

AFLUtils$e_merge(el, mode = "AND")

Arguments

el

A list of R expressions

mode

'AND' | 'OR'. Logical relationships when merging the expressions.

Returns

R expression