Base class of all ArrayOp classes

ArrayOp classes denote scidb array operations and operands, hence the name.

Details

One operation consists of an scidb operator and 1..* operands, of which the result can be used as an operand in another operation. Operands and Opreration results can all be denoted by ArrayOp.

Active bindings

dims

Dimension names

attrs

Attribute names

selected

Selected dimension and/or attribute names

dtypes

A named list, where key is dim/attr name and value is respective SciDB data type as string

dims_n_attrs

Dimension and attribute names

attrs_n_dims

Attribute and dimension names

is_schema_from_scidb

If the array schema is retrieved from SciDB or inferred locally in R

.private

For internal testing only. Do not access this field to avoid unintended consequences!!!

Methods

Public methods

ArrayOpBase$new()

ArrayOpBase$create_new()

ArrayOpBase$filter()

ArrayOpBase$select()

ArrayOpBase$reshape()

ArrayOpBase$change_schema()

ArrayOpBase$build_new()

ArrayOpBase$set_auto_fields()

ArrayOpBase$mutate()

ArrayOpBase$transmute()

ArrayOpBase$mutate_by()

ArrayOpBase$drop_dims()

ArrayOpBase$update()

ArrayOpBase$overwrite()

ArrayOpBase$spawn()

ArrayOpBase$limit()

ArrayOpBase$row_count()

ArrayOpBase$summarize_array()

ArrayOpBase$sync_schema()

ArrayOpBase$group_by()

ArrayOpBase$summarize()

ArrayOpBase$to_afl()

ArrayOpBase$to_schema_str()

ArrayOpBase$to_df_all()

ArrayOpBase$to_df()

ArrayOpBase$execute()

ArrayOpBase$versions()

ArrayOpBase$remove_versions()

ArrayOpBase$remove_self()

ArrayOpBase$inner_join()

ArrayOpBase$left_join()

ArrayOpBase$right_join()

ArrayOpBase$semi_join()

ArrayOpBase$persist()

ArrayOpBase$is_persistent()

ArrayOpBase$exists_persistent_array()

ArrayOpBase$array_meta_data()

ArrayOpBase$.set_meta()

ArrayOpBase$.get_meta()

ArrayOpBase$finalize()

Method new()

Return ArrayOp field types

NOTE: private$info has to be defined, otherwise returns NULL

Get dimension specifications

A dimension spec str is formatted as "lower bound : upper bound : overlap : chunk_length", as seen in scidb show(array_name) operator. All dimensions' data types are int64.

Validate fields existence according the 'type' which defaults to 'owned' fields

This function is useful for validating fields in use cases:

Generate afl for ArrayOp used in a join context (equi_join).

Create a new ArrayOp instance by joining with another ArrayOp

Currently implemented with scidb equi_join operator.

Create a new ArrayOp by matching a template against a source (self).

The result has the same schema with the source. All fields in the template are compared to their matching source fields by equality, except for thos in lower_bound/upper_bound which will be used as a range [lower_bound, upper_bound].

Create a new ArrayOp instance by loading a file and checking it against an ArrayOp template (self).

The ArrayOp instance where this function is called from serves as a template. By defulat, it assumes file column headers match the template's dims and attrs; otherwise an explicit file_headers can be provided and will be used to match the template's schema.

Create a new ArrayOp instance that with added auto incremented fields

Create a new ArrayOp instance that has an anti-collision field set according to a template arrayOp

The source (self) operand should have N fields given the target has N+1 dimensions. The one missing field is treated as the anti-collision field.

Return AFL suitable for retrieving data.frame.

scidb::iquery has a param only_attributes, which, if set TRUE, will effectively drop all dims.

Returns AFL when self used as an operand in another parent operation.

By default, 1. dimensions are not dropped in parent operation; 2. no intent to select fields

Base class initialize function, to be called in sub-class

Usage

ArrayOpBase$new(
  raw_afl,
  dims = as.character(c()),
  attrs = as.character(c()),
  dtypes = list(),
  dim_specs = list(),
  ...,
  metaList
)

Arguments

field_names

R character. If NULL, defaults to self$dims_n_attrs, ie. dimensions and attributes.

.raw

Default FALSE, full data types are returned; if set TRUE, only the raw data types are returned (raw data types are string, int32, int64, bool, etc, without scidb attribute specs such as: string compression 'zlib')

dim_names

Default NULL equals all dimensions.

keyFileds

Field names as join keys

keep_dimensiosn

If keep_dimensions is specified in scidb equi_join operator

right

The other ArrayOp instance to join with.

on_left

R character vector. Join keys from the left (self).

on_right

R character vector. Join keys from the right. Must be of the same length as on_left

on_both

Join keys on both operand.

.auto_select

default: FALSE. If set to TRUE, the resultant ArrayOp instance will auto select the fields that are selected in the left and right operands but not in the right join keys (since they are masked by equi_join operator).

settings

equi_join settings, a named list where both key and values are strings.

.dim_mode

How to reshape the resultant ArrayOp. Same meaning as in ArrayOp$reshape function. By default, dim_mode = 'keep', the artificial dimensions, namely instance_id and value_no from equi_join are retained. If set to 'drop', the artificial dimensions will be removed. See ArrayOp$reshape for more details.

.artificial_field

As in ArrayOp$reshape, it defaults to a random field name. It can be safely ignored in client code. It exists only for test purposes.

template

A data.frame or ArrayOp used to reduce the number of source cells without changing its schema

op_mode

'filter', 'cross_between'

lower_bound

Field names as lower bounds.

upper_bound

Field names as upper bounds.

field_mapping

A named list where name is source field name and value is template field name. Default NULL: fields are mapped between template and source by field names only. If there is mapping fields in the template which are intended for lower or upper bound, provide an empty list or a list with matching fields

filepath

A single file path

aio_settings

Customized settings of aio_input

field_conversion

If NULL (default), use template's field type to convert aio_input attributes; Or provide a list for customized field conversion

file_headers

Column headers of the input file regardless of whether there is a header line in the file. Default NULL assumes file headers match self$dims_n_attrs. If the headers order are different or there are columns to skip, please provide a string vector, in which case only columns with matching template field are loaded. Names of the unmatching column headers are irrelevant.

field_vec

A named string vector

reference

ArrayOp instance to draw existing max id from

source_field

ref_field

source_start

ref_start

new_field

target

A target arrayOp that the source draws anti-collision dimension from.

anti_collision_field

a target dimension name which exsits only to resolve cell collision (ie. cells with the same dimension coordinate).

drop_dims

Whether self's dimensions are dropped when generating AFL for data.frame conversion

drop_dims

Whether self dimensions will be dropped in parent operations

selected_fields

which fields are selected no matter what the parent operation is. If NULL, self fields will pass on by default depending on the parent operation.

info

A list that stores ArrayOp meta data, e.g. field types

Returns

a named list as field_names, where absent fields or fields without data types will cause error; unless .strict=F, absent fields are ignored

Method create_new()

Return a new arrayOp instance with the same version as self

Work in sub-class without requiring class names or constructor function.

Usage

ArrayOpBase$create_new(...)

Arguments

...

The samme params with Repo$ArrayOp(...)

Returns

A new arrayOp

Method filter()

Create a new ArrayOp instance by using a filter expression on the parent ArrayOp

Similar to SQL where clause and dplyr::filter.

Usage

ArrayOpBase$filter(
  ...,
  expr = NULL,
  missing_fields_error_template = NULL,
  regex_func = getOption("arrayop.regex_func", default = "regex"),
  ignore_case = getOption("arrayop.ignore_case", default = TRUE)
)

Arguments

missing_fields_error_template

Error template for missing fields. Only one %s is allowed which is substituted with an concatnation of the missing fields separated by commas.

regex_func

A string of regex function implementation. Due to scidb compatiblity issue with its dependencies, the regex function from boost library may not be available Supported values: rsub, regex

ignore_case

A Boolean. If TRUE, ignore case in string match patterns. Otherwise, perform case-sensitive regex matches.

Returns

A new arrayOp

Method select()

Create a new ArrayOp instance with selected fields

NOTE: this does NOT change the to_afl output, but explicitly state which field(s) are retained if used in a parent operation that changes its schema, e.g. equi_join or to_df(only_attributes = T)

Usage

ArrayOpBase$select(...)

Arguments

...

Which field(s) are retained during a schema-change operation

Returns

A new arrayOp

Method reshape()

Create a new ArrayOp instance with a different schema/shape

Usage

ArrayOpBase$reshape(
  select = NULL,
  dtypes = NULL,
  dim_mode = "keep",
  artificial_field = .random_attr_name(),
  .force_project = TRUE
)

Arguments

select

Which attributes to select or create. In dim_mode='keep', select must be a non-empty list, where named items are derived new attributes and unamed string values are existing dimensions/attributes. Dimensions, selected or not, are all retained in dim_mode='keep'. In dim_mode='drop', select can be NULL, which effectively select all source dimensions and attributes. Unselected dimensions will be discarded.

dtypes

a named list to provide field data types for newly derived fields

dim_mode

a string keep, drop. In the default 'keep' mode, only attributes can be selected. All dimensions are kept. In 'drop' mode, dimensions are first converted to attributes, then selected.

artificial_field

ONLY relevant when dim_mode='drop'. A field name used as the artificial dimension name in 'drop' dim_mode (internally used by unpack scidb operator). By default, a random string is generated.

Returns

A new arrayOp

Method change_schema()

Change self schema according to a template

This operation throws away any fields that do not exist in template while keeping the self's data of the matching fields. Implemented by scidb redimension operator, but it allows for partial-fields match if strict=F.

Usage

ArrayOpBase$change_schema(template, strict = TRUE, .setting = NULL)

Arguments

template

an arrayOp instance that determines the resultant schema.

strict

If TRUE(default), requires self has all the template fields.

.setting

a string vector, where each item will be appended to the redimension operand. E.g. .setting = c('false', 'cells_per_chunk: 1234') ==> redimension(source, template, false, cells_per_chunk: 1234)

Returns

A new arrayOp

Method build_new()

Create a new ArrayOp instance from 'build'ing a data.frame

All matching fields are built as attributes of the result ArrayOp. Build operator accepts compound attribute types, so the result may have something like "build(<aa:string not null, ...)"

Usage

ArrayOpBase$build_new(df, artificial_field = .random_attr_name())

Arguments

df

a data.frame, where all column names must all validate template fields.

artificial_field

A field name used as the artificial dimension name in build scidb operator By default, a random string is generated, and the dimension starts from 0. A customized dimension can be provided e.g. z=42:* or z=0:*:0:1000.

Returns

A new ArrayOp instance whose attributes share the same name and data types with the template's fields.

Method set_auto_fields()

Create a new ArrayOp instance that has auto incremented fields and/or anti-collision fields according to a template arrayOp

If the dimension count, attribute count and data types match between the source(self) and target, then no redimension will be performed, otherwise redimension on the source first.

Redimension mode requires all target fields exist on the source disregard of being attributes or dimensions. Redimension mode does not check on whether source data types match the target because auto data conversion occurs within scidb where necessary/applicable.

Usage

ArrayOpBase$set_auto_fields(
  target,
  source_auto_increment = NULL,
  target_auto_increment = NULL,
  anti_collision_field = NULL,
  join_setting = NULL,
  source_anti_collision_dim_spec = NULL
)

Arguments

target

A target ArrayOp the source data is written to.

source_auto_increment

a named number vector e.g. c(z=0), where the name is a source field and value is the starting index

target_auto_increment

a named number vector e.g. c(aid=0), where the name is a target field and value is the starting index. Here the target_auto_increment param only affects the initial load when the field is still null in the target array.

anti_collision_field

a target dimension name which exsits only to resolve cell collision (ie. cells with the same dimension coordinate).

Returns

A new arrayOp

Method mutate()

Usage

ArrayOpBase$mutate(..., .dots = NULL, .sync_schema = TRUE)

Method transmute()

Usage

ArrayOpBase$transmute(..., .dots = NULL, .sync_schema = TRUE)

Method mutate_by()

Usage

ArrayOpBase$mutate_by(
  data_array,
  keys = NULL,
  updated_fields = NULL,
  .redimension_setting = NULL,
  .join_setting = NULL
)

Method drop_dims()

Usage

ArrayOpBase$drop_dims(
  mode = "unpack",
  .chunk_size = NULL,
  .unpack_dim = dbutils$random_field_name()
)

Method update()

Update Target array with self's content

Similar behavior to scidb insert operator Fields of self and Target must match by raw data type. Field names are irrelevant.

Usage

ArrayOpBase$update(target)

Arguments

target

An arrayOp instance where self's content is written to.

Returns

A new arrayOp that encapsulates the insert operation

Method overwrite()

Overwrite Target array with self's content

Warning: Target's content will be erased and filled with self's content. Similar behavior to scidb store operator. Fields of self and Target must match by raw data type. Field names are irrelevant. See ArrayOp$change_schema

Usage

ArrayOpBase$overwrite(target)

Arguments

target

An arrayOp instance where self's content is written to.

Returns

A new arrayOp that encapsulates the insert operation

Method spawn()

Create a new ArrayOp instance from a self template

This function is mainly for array schema string generation where we might want to rename and/or exclude certain fields of the template. Data types and dimension specs will be inherited from the template unless provided.

Note: except for scidb build or aio_input operators, the spawned ArrayOp is not meaningful semantically. So do not use this function for operations other than 'build'/ArrayOp$build_new and 'aio_input'/ArrayOp$load_file

Usage

ArrayOpBase$spawn(
  afl_str = "spawned array_op (as template only)",
  renamed = NULL,
  added = NULL,
  excluded = NULL,
  dtypes = NULL,
  dim_specs = NULL
)

Arguments

renamed

Rename fields

added

New fields

excluded

Fields to exclude from self

dtypes

Data types for any existing or new fields, which default to self's data types if available.

dim_specs

A list of array dimension specs if dimensions are changed.

Returns

A new arrayOp instance

Method limit()

Get the first n rows of an array

Usage

ArrayOpBase$limit(n = 5, skip = NULL)

Arguments

n

How many rows to take

skip

How many rows to skip before taking

Method row_count()

Usage

ArrayOpBase$row_count()

Method summarize_array()

Usage

ArrayOpBase$summarize_array()

Method sync_schema()

Create a new array_op with actual schema from SciDB

Useful in confirmming the schema of transient arrays with complex operations If we are sure the array schema is already from SciDB, then just return self.

Usage

ArrayOpBase$sync_schema()

Method group_by()

Usage

ArrayOpBase$group_by(...)

Method summarize()

Usage

ArrayOpBase$summarize(..., .dots = NULL)

Method to_afl()

AFL representation of the ArrayOp instance

The ArrayOp instance may have 'selected' fields but they are not reflected in the result. 'selected' fields are meaningful in to_df_afl and .to_afl_explicit functions, where the parent operation treats dimension and attributes differently.

Usage

ArrayOpBase$to_afl()

Returns

an AFL expression string

Method to_schema_str()

Return a schema representation of the ArrayOp <attr1 , attr2 ...> [dim1 ;dim2]

Usage

ArrayOpBase$to_schema_str()

Returns

An AFL string

Method to_df_all()

Run query encapsulated by current array_op and return a R data.frame with all dimensions and attributes as columns.

Usage

ArrayOpBase$to_df_all()

Method to_df()

Run query encapsulated by current array_op and return a R data.frame with all attributes as columns.

Usage

ArrayOpBase$to_df()

Method execute()

Usage

ArrayOpBase$execute()

Method versions()

Usage

ArrayOpBase$versions()

Method remove_versions()

Usage

ArrayOpBase$remove_versions(version_id = NULL)

Method remove_self()

Usage

ArrayOpBase$remove_self()

Method inner_join()

Usage

ArrayOpBase$inner_join(
  right,
  on_left = NULL,
  on_right = NULL,
  on_both = NULL,
  left_alias = "_L",
  right_alias = "_R",
  join_mode = "equi_join",
  settings = NULL
)

Method left_join()

Usage

ArrayOpBase$left_join(
  right,
  on_left = NULL,
  on_right = NULL,
  on_both = NULL,
  left_alias = "_L",
  right_alias = "_R",
  settings = NULL
)

Method right_join()

Usage

ArrayOpBase$right_join(
  right,
  on_left = NULL,
  on_right = NULL,
  on_both = NULL,
  left_alias = "_L",
  right_alias = "_R",
  settings = NULL
)

Method semi_join()

Similar to dplyr::semi_join

params field_mapping, lower_bound and upper_bound, if provided, must be named list, where the.names are from the source array (i.e. self), and values are from the right operand df_or_arrayop

Usage

ArrayOpBase$semi_join(
  df_or_arrayop,
  field_mapping = NULL,
  lower_bound = NULL,
  upper_bound = NULL,
  mode = "auto",
  filter_threshold = 200L,
  upload_threshold = 6000L
)

Method persist()

Usage

ArrayOpBase$persist(save_array_name = NULL, .temp = FALSE, .gc = TRUE)

Method is_persistent()

Usage

ArrayOpBase$is_persistent()

Method exists_persistent_array()

Usage

ArrayOpBase$exists_persistent_array()

Method array_meta_data()

Usage

ArrayOpBase$array_meta_data()

Method .set_meta()

Set ArrayOp meta data directly

Useful in keeping reference of other relevant R objects that should have a same life cycle with the ArrayOp instance.

Usage

ArrayOpBase$.set_meta(key, value)

Arguments

key

A string key

value

A value of any type

Returns

NULL

Method .get_meta()

Get ArrayOp meta data directly

Usage

ArrayOpBase$.get_meta(key)

Arguments

key

A string key

Returns

An metadata object of any type registered with key

Method finalize()

Usage

ArrayOpBase$finalize()