Base class of all ArrayOp classes

ArrayOp classes denote scidb array operations and operands, hence the name.

Details

One operation consists of an scidb operator and 1..* operands, of which the result can be used as an operand in another operation. Operands and Opreration results can all be denoted by ArrayOp.

Super class

arrayop::ArrayOpBase -> ArrayOpV19

Methods

Public methods

ArrayOpV19$clone()

Inherited methods

arrayop::ArrayOpBase$.get_meta()

arrayop::ArrayOpBase$.set_meta()

arrayop::ArrayOpBase$.to_afl_explicit()

arrayop::ArrayOpBase$.to_join_operand_afl()

arrayop::ArrayOpBase$build_new()

arrayop::ArrayOpBase$change_schema()

arrayop::ArrayOpBase$create_array_cmd()

arrayop::ArrayOpBase$create_new()

arrayop::ArrayOpBase$create_new_with_same_schema()

arrayop::ArrayOpBase$get_absent_fields()

arrayop::ArrayOpBase$get_dim_specs()

arrayop::ArrayOpBase$get_field_types()

arrayop::ArrayOpBase$initialize()

arrayop::ArrayOpBase$join()

arrayop::ArrayOpBase$load_file()

arrayop::ArrayOpBase$match()

arrayop::ArrayOpBase$mutate()

arrayop::ArrayOpBase$overwrite()

arrayop::ArrayOpBase$remove_array_cmd()

arrayop::ArrayOpBase$remove_array_versions_cmd()

arrayop::ArrayOpBase$reshape()

arrayop::ArrayOpBase$select()

arrayop::ArrayOpBase$set_anti_collision_field()

arrayop::ArrayOpBase$set_auto_fields()

arrayop::ArrayOpBase$set_auto_increment_field()

arrayop::ArrayOpBase$set_null_fields()

arrayop::ArrayOpBase$spawn()

arrayop::ArrayOpBase$to_afl()

arrayop::ArrayOpBase$to_df_afl()

arrayop::ArrayOpBase$to_schema_str()

arrayop::ArrayOpBase$update()

arrayop::ArrayOpBase$where()

arrayop::ArrayOpBase$write_to()

Method clone()

The objects of this class are cloneable with this method.

Usage

ArrayOpV19$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.