RepoBase class manages ArrayOp objects in one scidb installation

A repository for scidb arrays and derived ArrayOp objects that reside in one scidb installation

Details

RepoBase is the base class of scidb-version-specific Repo classes. Common ogic/behaviors across all supported scidb versions are placed here. Version-specific logic/behaviors should be implemented in RepoBase sub-classes.

Active bindings

.private

For internal testing only. Do not access this field to avoid unintended consequences!!!

setting_debug

Whether to enable debug mode. Default FALSE. Repo$setting_debug to get current setting; or Repo$setting_debug = T to change current setting.

setting_use_aio_input

Whether to use aio_input for data frame uploading. Default FALSE. Repo$setting_use_aio_input to get current setting; or Repo$setting_use_aio_input = T to change current setting.

Methods

Public methods

Repo$new()

Repo$query()

Repo$execute()

Repo$get_array()

Repo$register_array()

Repo$load_arrayop_from_scidb()

Repo$load_arrayops_from_scidb_namespace()

Repo$load_arrayops_from_config()

Repo$upload_df()

Repo$save_as_array()

Repo$.create_array()

Repo$.remove_array_versions()

Repo$.remove_array()

Repo$nrow()

Repo$limit()

Repo$clone()

Method new()

Initialize function.

Create a new Repo instance

Do NOT call this initialize function directly. Call newRepo instead to get a new Repo instance.

Usage

Repo$new(dependency_object = NULL)

Method query()

Run a query and get a data frame

The query statement stmt is a string.

Usage

Repo$query(what, ..., .dry_run = FALSE, .raw = TRUE, .env = parent.frame())

Arguments

...

Arguments passed directly to scidb::iquery function

.dry_run

If TRUE, only return evaluated query statement. Default FALSE

.raw

If TRUE and what is a string, do not evaluate, otherwise evaluate what. Default FALSE. Only applicable if what is a string.

.env

An R env or list for R variable substitution

stmt

A string, where

array aliases are replaced by corresponding arrayOp instances

functions following arrays are called as arrayOp functions

extra arrayOp functions can be provided in parentheses

any other variables in .env will be replaced by its values

Returns

A data frame

Method execute()

Execute a query. No result is returned.

The query statement stmt is a string. No result is returned.

Usage

Repo$execute(what, ..., .dry_run = FALSE, .raw = TRUE, .env = parent.frame())

Arguments

what

A string, where

array aliases are replaced by corresponding arrayOp instances

functions following arrays are called as arrayOp functions

extra arrayOp functions can be provided in parentheses

any other variables in .env will be replaced by its values

...

Arguments passed directly to scidb::iquery function

.dry_run

If TRUE, only return evaluated query statement. Default FALSE

.raw

If TRUE and what is a string, do not evaluate, otherwise evaluate what. Default FALSE. Only applicable if what is a string.

.env

An R env or list for R variable substitution

Returns

NULL

Method get_array()

Get an ArrayOp instance

Param what can be a string or arrayOp instance. If a string, then what is either a registered array alias, or raw array schema (e.g. myArray <a:int32, b:string> [d=0::0:])

This function is provide unified access of arrayOp instances

Usage

Repo$get_array(what)

Arguments

what

An array alias, raw array schema or arrayOp instance

Returns

An ArrayOp instance

Method register_array()

Register a list of arrays with aliases

Register a list of arrays which can be later accessed via their aliases

Usage

Repo$register_array(items)

Arguments

items

A list where names are aliases and values are arrayOp instances or raw array names

Method load_arrayop_from_scidb()

Get an ArrayOp instance from scidb by its full name

No cache is provided with this function. See register_array for details on registering an array alias with cache.

Usage

Repo$load_arrayop_from_scidb(full_array_name)

Arguments

full_array_name

A fully qualified array name (e.g. myNamespace.arrayName)

Returns

An ArrayOp instance

Method load_arrayops_from_scidb_namespace()

Get a list of arrays from a scidb namespace

Usage

Repo$load_arrayops_from_scidb_namespace(namespace)

Arguments

namespace

A scidb namespace

Returns

A list of ArrayOp instances where names are the array names (without namespace) and values are the arrayOp instances whose to_afl() function return fully qualified array names

Method load_arrayops_from_config()

Load a list of ArrayOp instances from a config object

Only load and return a list of arrayOp instances. Repo's array registery will not be updated. If desired, you can loaded = repo$load_arrayops_from_config; repo$register_array(loaded) to register loaded arrays.

Usage

Repo$load_arrayops_from_config(config)

Arguments

config

A config object (R list) normally loaded from a yaml config file. The config list must have 'namespace' and 'arrays' keys. Value of 'namespace' is a single string for the default namespace. Value of 'arrays' is a list of unamed elements, where each element is a list of three items:

'alias': a single string for array alias

'name': raw array name. If without a namespace, then the default namespace is applied. Otherwise, if in 'namespace.rawArrayName', then use it directly disregard of the default namespace.

'schema': a single string for array schema. E.g. "<aa:string, b:bool null> [da=0::0:]

Returns

A list of ArrayOp instances where names are array aliases and values are the ArrayOp instances

Method upload_df()

Upload a data frame to scidb

Upload a data frame into scidb with data types from the template. The data frame can have fewer columns than the template, but it can NOT have columns absent in the template.

Usage

Repo$upload_df(
  df,
  template,
  temp = TRUE,
  use_aio_input = setting_use_aio_input,
  ...
)

Arguments

df

a data frame whose column names has been sanitized

template

An array alias, schema string, arrayOp; or a named list of data types

temp

Whether to upload the df as a temporary array. Default TRUE.

use_aio_input

Whether to use aio_input for upload. Default FALSE. Same meaning as in scidb::as.scidb

...

Other args passed to scidb::as.scidb, e.g. gc = TRUE

Returns

An arrayOp instance with uploaded data. The resultant arrayOp$to_afl() is the temporary array name R_array****, unless specified explicitly with name='newName' The resultant arrayOp schema is consisted of 1. artificial dimension(s) from upload or aio_input; 2. attributes whose names are the data frame column names and data types are those in the template.

Method save_as_array()

Save an arrayOp or AFL statement as a scidb array wrapped as a new arrayOp

No all things can be done in one scidb operation, sometimes we need to store intermediate results as arrays. This is when we need to save materialized arrayOp into a new array, which is then wrapped in a new arrayOp instance. To get the new array's raw name, call the result arrayOp$to_afl() function.

The ... params are passed in to scidb::store. We can specify an array name name='myArray'(default:a random name in public namespace), create a temporary array temp=T(default:F), or suppress auto gc gc=F(default:T). Refer to scidb documentation for more details.

Usage

Repo$save_as_array(what, ...)

Arguments

what

an arrayOp or an AFL statement (string)

...

Same params as used in scidb::store function, e.g. temp = F, gc = T, name = 'myArray'

Returns

An arrayOp whose to_afl() is the newly created array name

Method .create_array()

Create a new array using the operand's schema

Usage

Repo$.create_array(array_or_alias, new_array_name = NULL)

Arguments

array_or_alias

The array operand as a template, either an arrayOp isntance or a registered array alias

new_array_name

The new array name. If NULL, it uses the operand's to_afl() result as the new name.

Method .remove_array_versions()

Remove array versions

Usage

Repo$.remove_array_versions(array_or_alias, ...)

Arguments

array_or_alias

The array operand as a template, either an arrayOp isntance or a registered array alias

Method .remove_array()

Remove array. Careful: Cannot be undone!!!

Usage

Repo$.remove_array(array_or_alias)

Arguments

array_or_alias

The array operand as a template, either an arrayOp isntance or a registered array alias

Method nrow()

Get the number of rows of an arrayOp

Usage

Repo$nrow(what, .raw = TRUE, .env = parent.frame())

Arguments

what

An arrayOp, raw AFL or R expression string

.raw

If TRUE, and what is string, treat it as raw AFL. Otherwise parse what as R expression

.env

The env where what is evaluated as R expression. Only applicable when what is R expression string and .raw = F

Method limit()

Get the first count rows of an array

Usage

Repo$limit(what, count, offset = NULL, ..., .raw = TRUE, .env = parent.frame())

Arguments

what

An arrayOp, raw AFL or R expression string

count

How many rows to take

offset

How many rows to skip before taking

...

Arguments passed directly to scidb::iquery function

.raw

If TRUE, and what is string, treat it as raw AFL. Otherwise parse what as R expression

.env

The env where what is evaluated as R expression. Only applicable when what is R expression string and .raw = F

Method clone()

The objects of this class are cloneable with this method.

Usage

Repo$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.