Create an array_op from a schema string with an optional array name

Create an array_op from a schema string with an optional array name

Create an array_op from a schema string with an optional array name

Methods

Public methods

ScidbConnection$new()

ScidbConnection$is_connected()

ScidbConnection$connect()

ScidbConnection$scidb_version()

ScidbConnection$conn_args()

ScidbConnection$query_all()

ScidbConnection$query()

ScidbConnection$execute()

ScidbConnection$create_array()

ScidbConnection$array()

ScidbConnection$array_from_schema()

ScidbConnection$array_from_afl()

ScidbConnection$array_from_stored_afl()

ScidbConnection$array_from_df()

ScidbConnection$upload_df()

ScidbConnection$compile_df()

ScidbConnection$fread()

ScidbConnection$clone()

Method new()

Usage

ScidbConnection$new()

Arguments

array_like

An array_op, schema string or existing scidb array name

Method is_connected()

Usage

ScidbConnection$is_connected()

Method connect()

Usage

ScidbConnection$connect(connection_args = NULL)

Method scidb_version()

Usage

ScidbConnection$scidb_version()

Method conn_args()

Usage

ScidbConnection$conn_args()

Method query_all()

Usage

ScidbConnection$query_all(afl_str)

Method query()

Usage

ScidbConnection$query(afl_str)

Method execute()

Usage

ScidbConnection$execute(afl_str)

Method create_array()

Usage

ScidbConnection$create_array(name, schema_template, .temp = FALSE)

Method array()

Usage

ScidbConnection$array(array_name)

Method array_from_schema()

Usage

ScidbConnection$array_from_schema(schema_str)

Method array_from_afl()

Usage

ScidbConnection$array_from_afl(afl_str)

Method array_from_stored_afl()

Create a persistent array_op by storing AFL as an array

Usage

ScidbConnection$array_from_stored_afl(
  afl_str,
  save_array_name = NULL,
  .temp = FALSE,
  .gc = TRUE
)

Arguments

.temp

Whether to create a temporary scidb array. Only effective when save_array_name = NULL

Method array_from_df()

Usage

ScidbConnection$array_from_df(
  df,
  template = NULL,
  build_or_upload_threshold = 8000L,
  build_dim_spec = .random_field_name(),
  skip_scidb_schema_check = FALSE,
  force_template_schema = FALSE,
  ...
)

Method upload_df()

Usage

ScidbConnection$upload_df(
  df,
  template = NULL,
  name = dbutils$random_array_name(),
  force_template_schema = FALSE,
  upload_by_vector = FALSE,
  .use_aio_input = FALSE,
  .temp = FALSE,
  .gc = TRUE
)

Method compile_df()

Usage

ScidbConnection$compile_df(
  df,
  template = NULL,
  build_dim_spec = dbutils$random_field_name(),
  force_template_schema = FALSE,
  skip_scidb_schema_check = FALSE
)

Method fread()

Get a transient array_op by loading a file with aio_input

Mimic data.table::fread

Usage

ScidbConnection$fread(
  file_path,
  template = NULL,
  header = TRUE,
  sep = "\t",
  col.names = NULL,
  mutate_fields = NULL,
  auto_dcast = FALSE,
  nrow = 10L,
  instances = NULL,
  .aio_settings = NULL
)

Arguments

auto_dcast

logical, default F. If TRUE, all non-string fields are dcast'ed with `dcast(ax, int64(null)), where ax is the 0-indexed mapping attribute name (e.g. a0, a1, etc), and int64 is the template field type

Method clone()

The objects of this class are cloneable with this method.

Usage

ScidbConnection$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.