Create AFL expressions from R expressions

This is a convenience function for AFL generation.

afl(..., envir = parent.frame())

Arguments

...
In the ellipsis arg, any R functions right after a pipe sign `|`` is converted to a scidb operator of the same name. All regular functions are first evaluated in the calling environment, and then convereted to strings depending on the result types. ArrayOp => ArrayOp$to_afl(), v:NonEmptyVector => paste(v, collapse=','), NULL is ignored.

envir
The environment where expressions are evaluated. Default: the calling env.

Value

AFL string

Details

Any a | op_name(b) call will be translated to op_name(a, b) in R, then translated to AFL: op_name(a, b)

Using this syntax, we can chain multiple AFL operators

E.g. 'array' | filter('a > 3 and b < 4') | project('a', 'b') will be translated into: project(filter(array, a > 3 and b < 4), 'a', 'b') Use NULL if no 2nd operand is needed. E.g. 'array' | op_count => op_count(array)