Create a scidb-version-specific Repo instance

The created Repo instance is automatically selected based on Repo-Scidb compatibility

newRepo(dependency_obj = NULL, db = NULL)

Arguments

dependency_obj
A named list that contains 4 functions: get_scidb_version, query, execute and get_schema_df If left NULL, will require a scidb connection 'db' to create the dependency_obj automatically.

db
a scidb connection returned by scidb::scidbconnect. If provided, a dependency_obj is automatically created with default implementations.

Details

This should be only way to create a new Repo instance, unless a specific Repo version is known/desired beforehand, which is useful in test cases.