Base class of all ArrayOp classes

ArrayOp classes denote scidb array operations and operands, hence the name.

Details

One operation consists of an scidb operator and 1..* operands, of which the result can be used as an operand in another operation. Operands and Opreration results can all be denoted by ArrayOp.

Active bindings

dims

Dimension names

attrs

Attribute names

selected

Selected dimension and/or attribute names

dtypes

A named list, where key is dim/attr name and value is respective SciDB data type as string

raw_dtypes

A named list, where key is dim/attr name and value is first part of respective SciDB data type as string

dims_n_attrs

Dimension and attribute names

attrs_n_dims

Attribute and dimension names

is_schema_from_scidb

If the array schema is retrieved from SciDB or inferred locally in R

is_scidb_data_frame

Whether current array_op is a regular array or SciDB data frame (array with hidden dimensions; not to be confused with R data frames)

.private

For internal testing only. Do not access this field to avoid unintended consequences!!!

Methods

Public methods

ArrayOpBase$new()

ArrayOpBase$filter()

ArrayOpBase$select()

ArrayOpBase$change_schema()

ArrayOpBase$set_auto_fields()

ArrayOpBase$mutate()

ArrayOpBase$transmute()

ArrayOpBase$mutate_by()

ArrayOpBase$drop_dims()

ArrayOpBase$update()

ArrayOpBase$overwrite()

ArrayOpBase$spawn()

ArrayOpBase$limit()

ArrayOpBase$cell_count()

ArrayOpBase$summarize_array()

ArrayOpBase$sync_schema()

ArrayOpBase$group_by()

ArrayOpBase$summarize()

ArrayOpBase$to_afl()

ArrayOpBase$to_schema_str()

ArrayOpBase$to_df_all()

ArrayOpBase$to_df()

ArrayOpBase$execute()

ArrayOpBase$versions()

ArrayOpBase$remove_versions()

ArrayOpBase$remove_array()

ArrayOpBase$inner_join()

ArrayOpBase$left_join()

ArrayOpBase$right_join()

ArrayOpBase$semi_join()

ArrayOpBase$persist()

ArrayOpBase$is_persistent()

ArrayOpBase$exists_persistent_array()

ArrayOpBase$array_meta_data()

ArrayOpBase$finalize()

Method new()

Return ArrayOp field types as a list

Get dimension specifications as a list

A dimension spec str is formatted as "lower bound : upper bound : overlap : chunk_length", as seen in scidb show(array_name) operator. All dimensions' data types are int64, as returned by get_field_types function.

Generate afl for ArrayOp used in a join context (equi_join).

Create a new ArrayOp instance by joining with another ArrayOp

Currently implemented with scidb equi_join operator.

Disambiguate selected fields of join

If no name clashes between selected fields of both operands, nothing changes. Otherwise, add suffix to duplicated fields for each operand. E.g. if a field named 'value' exists on both operands and is retained, it will be renamed as 'value_left' and 'value_right' if left_alias and right_alias are '_left' and '_right', respectively.

Create a new ArrayOp by matching a template against a source (self).

The result has the same schema with the source. All fields in the template are compared to their matching source fields by equality, except for thos in lower_bound/upper_bound which will be used as a range [lower_bound, upper_bound].

Return a new arrayOp instance with the same version as self

Work in sub-class without requiring class names or constructor function.

Create a new ArrayOp instance from 'build'ing a data.frame

All matching fields are built as attributes of the result ArrayOp. Build operator accepts compound attribute types, so the result may have something like "build(<aa:string not null, ...)"

Create a new ArrayOp instance by loading a file and checking it against an ArrayOp template (self).

The ArrayOp instance where this function is called from serves as a template. By defulat, it assumes file column headers match the template's dims and attrs; otherwise an explicit file_headers can be provided and will be used to match the template's schema.

Create a new ArrayOp instance that with added auto incremented fields

Create a new ArrayOp instance that has an anti-collision field set according to a template arrayOp

The source (self) operand should have N fields given the target has N+1 dimensions. The one missing field is treated as the anti-collision field.

Return AFL suitable for retrieving data.frame.

scidb::iquery has a param only_attributes, which, if set TRUE, will effectively drop all dims.

Returns AFL when self used as an operand in another parent operation.

By default, 1. dimensions are not dropped in parent operation; 2. no intent to select fields

Base class initialize function, to be called in sub-class internally.

Always use ScidbConnection to get array_op instances.

Usage

ArrayOpBase$new(
  raw_afl,
  dims = as.character(c()),
  attrs = as.character(c()),
  dtypes = list(),
  dim_specs = list(),
  ...,
  meta_list
)

Arguments

dims

A string vector used as dimension names

attrs

A string vector used as attribute names

dtypes

A named list of strings, where names are attribute names and values are full scidb data types. E.g. dtypes = list(field_str = "string NOT NULL", field_int32 = "int32")

dim_specs

A named list of string, where names are dimension names and values are dimension specs. E.g. dim_sepcs = list(da = "0:*:0:*", chrom = "1:24:0:1").

...

The samme params with Repo$ArrayOp(...)

...

A named list of metadata items, where names are used as keys in private$set_meta and private$get_meta functions.

meta_list

A list that stores ArrayOp meta data, e.g. field types If provided, other regular parms are not allowed.

field_names

R character. If NULL, defaults to self$dims_n_attrs, ie. dimensions and attributes.

.raw

Default FALSE, full data types are returned; if set TRUE, only the raw data types are returned (raw data types are string, int32, int64, bool, etc, without scidb attribute specs such as: string compression 'zlib')

dim_names

Default NULL equals all dimensions.

keyFileds

Field names as join keys

keep_dimensiosn

If keep_dimensions is specified in scidb equi_join operator

right

The other ArrayOp instance to join with.

on_left

R character vector. Join keys from the left (self).

on_right

R character vector. Join keys from the right. Must be of the same length as on_left

on_both

Join keys on both operand.

settings

equi_join settings, a named list where both key and values are strings.

.dim_mode

How to reshape the resultant ArrayOp. Same meaning as in ArrayOp$reshape function. By default, dim_mode = 'keep', the artificial dimensions, namely instance_id and value_no from equi_join are retained. If set to 'drop', the artificial dimensions will be removed. See ArrayOp$reshape for more details.

.artificial_field

As in ArrayOp$reshape, it defaults to a random field name. It can be safely ignored in client code. It exists only for test purposes.

template

A data.frame or ArrayOp used to reduce the number of source cells without changing its schema

op_mode

'filter', 'cross_between'

lower_bound

Field names as lower bounds.

upper_bound

Field names as upper bounds.

field_mapping

A named list where name is source field name and value is template field name. Default NULL: fields are mapped between template and source by field names only. If there is mapping fields in the template which are intended for lower or upper bound, provide an empty list or a list with matching fields

df

a data.frame, where all column names must all validate template fields.

artificial_field

A field name used as the artificial dimension name in build scidb operator By default, a random string is generated, and the dimension starts from 0. A customized dimension can be provided e.g. z=42:* or z=0:*:0:1000.

filepath

A single file path

aio_settings

Customized settings of aio_input

field_conversion

If NULL (default), use template's field type to convert aio_input attributes; Or provide a list for customized field conversion

file_headers

Column headers of the input file regardless of whether there is a header line in the file. Default NULL assumes file headers match self$dims_n_attrs. If the headers order are different or there are columns to skip, please provide a string vector, in which case only columns with matching template field are loaded. Names of the unmatching column headers are irrelevant.

field_vec

A named string vector

reference

ArrayOp instance to draw existing max id from

source_field

ref_field

source_start

ref_start

new_field

target

A target arrayOp that the source draws anti-collision dimension from.

anti_collision_field

a target dimension name which exsits only to resolve cell collision (ie. cells with the same dimension coordinate).

drop_dims

Whether self's dimensions are dropped when generating AFL for data.frame conversion

drop_dims

Whether self dimensions will be dropped in parent operations

selected_fields

which fields are selected no matter what the parent operation is. If NULL, self fields will pass on by default depending on the parent operation.

Returns

a named list where keys are field names, and values are scidb data types (single word or multiple words).

Throw an error if field_names are not all valid.

Method filter()

Create a new ArrayOp instance with filter expressions

Similar to dplyr::filter, fields are not quoted.

Operators for any type of fields include ==, !=, %in%, %not_in%. To test whether a field is null, use unary operators: is_null, not_null.

Special binary operators for string fields include: %contains%, %starts_with%, %ends_with%, %like%, where only %like% takes a regular expression and other operators escape any special characters in the right operand.

Operators for numeric fields include: >, <, >=, <=

Usage

ArrayOpBase$filter(
  ...,
  .expr = NULL,
  .validate_fields = TRUE,
  .regex_func = getOption("arrayop.regex_func", default = "regex"),
  .ignore_case = getOption("arrayop.ignore_case", default = TRUE)
)

Arguments

...

Filter expression(s) in R syntax. These expression(s) are not evaluated in R but first captured then converted to scidb expressions with appropriate syntax.

.expr

A single R expression, or a list of R exprs, or NULL. If provided, ... is ignored. Mutiple exprs are joined by 'and'. This param is useful when we want to pass an already captured R expression.

regex_func

A string of regex function implementation, default 'regex'. Due to scidb compatiblity issue with its dependencies, the regex function from boost library may not be available Currently supported options include 'rsub', and 'regex'

ignore_case

A Boolean, default TRUE. If TRUE, ignore case in string match patterns. Otherwise, perform case-sensitive regex matches.

Returns

A new arrayOp

Method select()

Create a new ArrayOp instance with selected fields

NOTE: this does NOT change the to_afl output, but explicitly state which field(s) should be retained if used in a parent operation that changes its schema, e.g. inner_join, left_join, right_join and to_df.

The selected fields are passed on to derived ArrayOp instances.

In all join operations, if no field is explicitly selected, then all fields are assumed be retained. In to_df, if no field is explicitly selected, only the attributes are retrieved as data frame columns. In to_df_all, if no field is explicitly selected, it is equivalent to select all dimensions and attributes.

Usage

ArrayOpBase$select(...)

Arguments

...

Which field(s) to retain.

Returns

A new arrayOp

Method change_schema()

Create a new ArrayOp instance whose schema is the same as the template.

This operation throws away any fields that do not exist in template while keeping the self's data of the matching fields.

Implemented by scidb redimension operator, but it allows for partial-fields match if strict=F.

Usage

ArrayOpBase$change_schema(template, strict = TRUE, .setting = NULL)

Arguments

template

an ArrayOp instance as the schema template.

strict

If TRUE(default), requires self has all the template fields.

.setting

a string vector, where each item will be appended to the redimension operand. E.g. .setting = c('false', 'cells_per_chunk: 1234') ==> redimension(source, template, false, cells_per_chunk: 1234)

Returns

A new arrayOp instance

Method set_auto_fields()

Create a new ArrayOp instance that has auto incremented fields and/or anti-collision fields according to a template arrayOp

If the dimension count, attribute count and data types match between the source(self) and target, then no redimension will be performed, otherwise redimension on the source first.

Redimension mode requires all target fields exist on the source disregard of being attributes or dimensions. Redimension mode does not check on whether source data types match the target because auto data conversion occurs within scidb where necessary/applicable.

Usage

ArrayOpBase$set_auto_fields(
  target,
  source_auto_increment = NULL,
  target_auto_increment = NULL,
  anti_collision_field = NULL,
  join_setting = NULL,
  source_anti_collision_dim_spec = NULL
)

Arguments

target

A target ArrayOp the source data is written to.

source_auto_increment

A single named integer, a single string, or NULL. Eg. c(z=0) for field 'z' in the source (ie. self) starting from 0; or a single string 'z' equivalent to c(z=0). If NULL, assume it to be the only dimension in self, normally from an artificial dimension of a build literal or unpack operation.

target_auto_increment

a named number vector or string vector or NULL. where the name is a target field and value is the starting index. E.g. c(aid=0, bid=1) means to set auto fields 'aid', 'bid' according to the target fields of the same name. If 'target' doesn't have a cell, then default values start from 0 and 1 for aid and bid, respectively. A string vector c("aid", "bid") is equvilant to c(aid=0, bid=0). NULL means treat all missing fields (absent in self but present in target) as 0-based auto increment fields. Here the target_auto_increment param only affects the initial load when the field is still null in the target array.

anti_collision_field

a target dimension name which exsits only to resolve cell collision (ie. cells with the same dimension coordinate).

Returns

A new arrayop instance

Method mutate()

Create a new ArrayOp instance with mutated fields

Similar to dplyr::mutate, fields of source (self) can be removed or added to the result arrayOp Any field that are not in the mutate expressions remain unchanged.

Usage

ArrayOpBase$mutate(..., .dots = NULL, .sync_schema = TRUE)

Arguments

...

Named R expressions. Names are field names in the result arrayOp and must not be empty. Set field = NULL to remove existing fields. E.g. abcd = NULL, def = def removes field 'abcd' and keep field 'def'.Values are R expressions similar to the filter method. E.g. a = b + 2, name = first + "-" + last, chrom = if(chrom == 'x') 23 else if(chrom == 'y') 24 else chrom

.dots

A named list of R expressions or NULL. If provided, the ... param is ignored. Useful when an a list of mutation expressions is already created and can be passed around.

.sync_schema

Whether to get the exact schema from scidb. Default TRUE will cause a scidb query to get the schema. Set to FALSE to avoid schema checking.

Returns

a new ArrayOp instance

Method transmute()

Create a new ArrayOp instance with mutated fields

Similar to dplyr::transmute, only listed fields are retained in the result arrayOp NOTE: Any field that are not in the mutate expressions will be discarded.

Usage

ArrayOpBase$transmute(..., .dots = NULL, .sync_schema = TRUE)

Arguments

...

R expressions. Names are optional. For each named expression, the name is used as field name in the result arrayOp. Unnamed expressions must be existing field names, unquoted, which result in unchanged source fields of self.Values are R expressions similar to the filter method. E.g. a = b + 2, name = first + "-" + last, chrom = if(chrom == 'x') 23 else if(chrom == 'y') 24 else chrom

.dots

A named list of R expressions or NULL. If provided, the ... param is ignored. Useful when an a list of mutation expressions is already created and can be passed around.

.sync_schema

Whether to get the exact schema from scidb. Default TRUE will cause a scidb query to get the schema. Set to FALSE to avoid schema checking.

Returns

a new ArrayOp instance

Method mutate_by()

Create a ArrayOp instance with the same schema of self, but different cells from 'data_array' for the 'updated_fields'.

Usage

ArrayOpBase$mutate_by(
  data_array,
  keys = NULL,
  updated_fields = NULL,
  .redimension_setting = NULL,
  .join_setting = NULL
)

Arguments

data_array

An ArrayOp instance that have at least two overlapping fields with self.

keys

Field names in both self and data_array. Cell content of these fields are from the 'self' arrayOp rather than 'data_array'.

updated_fields

Field names in both self and data_array. Cell content of these fields are from the 'data_array', NOT 'self'.

.redimension_setting

A list of strings used as the settings of scidb 'redimension' operator. Only applicable when a 'redimension' is needed.

.join_setting

A list of strings used as the settings of scidb 'join' operator. Only applicable when a 'join' is needed.

Returns

A new arrayOp with the same schema as self

Method drop_dims()

Create a new arrayOp by dropping dimensions of 'self'.

Use mode = 'unpack' to still keep an artificial dimension in result arrayOp. The dimension is 0-based, auto-incremented up until self$cell_count() - 1. 'unpack' mode is useful in taking advantage of this artifical dimension to auto populate other fields, e.g. in set_auto_fields.

Use mode = 'flatten' to return a scidb data frame which has no explicit dimensions.

Result arrayOp in both modes has attributes of self's attributes and dimensions.

Usage

ArrayOpBase$drop_dims(
  mode = "unpack",
  .chunk_size = NULL,
  .unpack_dim = dbutils$random_field_name()
)

Arguments

mode

String 'unpack' (default) or 'flatten'.

.chunk_size

NULL or an integer. Converted to the 'chunk_size' param in 'unpack' mode; and 'cells_per_chunk' in 'flatten' mode.

Method update()

Update the target array with self's content

Similar behavior to scidb insert operator. Require numbers of attributes and dimensions of self and target arrays match. Field names are irrelevant.

This function only returns an arrayOp with the update operation AFL encapsulated. No real action is performed in scidb until source$update(target)$execute() is called.

Usage

ArrayOpBase$update(target)

Arguments

target

An arrayOp instance where self's content is updated. Must be a persistent array, since it is meanlingless to update an array operation.

Returns

A new arrayOp that encapsulates the update operation

Method overwrite()

Overwrite the target array with self's content

Similar behavior to scidb store operator. Require numbers of attributes and dimensions of self and target arrays match. Field names are irrelevant.

This function only returns an arrayOp with the update operation AFL encapsulated. No real action is performed in scidb until source$overwrite(target)$execute() is called.

Warning: Target's content will be erased and filled with self's content.

Usage

ArrayOpBase$overwrite(target)

Arguments

target

An arrayOp instance where self's content is written to. Must be a persitent array either preexist or does not exist in scidb.

Returns

A new arrayOp that encapsulates the overwrite operation

Method spawn()

Create a new ArrayOp instance using 'self' as a template

This function is mainly for array schema string generation when we want to rename, add, and/or exclude certain fields of self, but still keep other unspecified fields unchanged.

Data types and dimension specs of existing fields are inherited from 'self' unless provided explicitly. New field data types default to NAs unless provided explicitly.

This function is normally used internally for arrayOp generation.

Usage

ArrayOpBase$spawn(
  afl_str = "spawned array_op (as template only)",
  renamed = NULL,
  added = NULL,
  excluded = NULL,
  dtypes = NULL,
  dim_specs = NULL
)

Arguments

afl_str

An AFL expression. In case of using the spawned result as a schema template only, the afl_str does not need to be provided. Otherwise, it should conform with the actual resultant arrayOp instance, which is very rare.

renamed

A list of renamed fields where names are old fields and values are new field names.

added

New fields added to result arrayOp. String vector or NULL.

excluded

Fields excluded from self. String vector or NULL.

dtypes

NULL or a named list of data types for fields of the result arrayOp, where names are field names, values (strings) are data types.

dim_specs

NULL or a named list of array dimension specs, where names are dimension names, values (strings) are dimension specs in scidb format.

Returns

A new arrayOp instance

Method limit()

Create a new arrayOp that encapsulate AFL for the first n cells of 'self'

We still need to append a to_df call to download the result as data frame.

Usage

ArrayOpBase$limit(n = 5, skip = NULL)

Arguments

n

How many cells to take

skip

How many rows to skip before taking

Returns

A new ArrayOp instance

Method cell_count()

Return the number of cells of 'self'

Usage

ArrayOpBase$cell_count()

Returns

A number of cells if the AFL that 'self' encapsulates is run.

Method summarize_array()

Return a data frame of the summary of the 'self' array

Implemented by scidb 'summarize' operator

Usage

ArrayOpBase$summarize_array(by_attribute = FALSE, by_instance = FALSE)

Arguments

by_attribute

Summarize by array attributes

by_instance

Summarize by array scidb instances

return

A data frame of the 'self' array summary

Method sync_schema()

Create a new arrayOp with actual schema from SciDB or 'self' if self$is_schema_from_scidb == T.

Useful in confirmming the schema of complex array operations. If the array schema is already retrieved from SciDB, then just return self.

Usage

ArrayOpBase$sync_schema()

Returns

An arrayOp instance

Method group_by()

Create a new arrayOp with 'group by' fields

The result arrayOp is identical to self except for the 'group_by' fields. Useful for summarize function, which will be converted into grouped_aggregate operations.

Usage

ArrayOpBase$group_by(...)

Returns

An arrayOp instance with group_by fields

Method summarize()

Create a new arrayOp with aggregated fields

Usage

ArrayOpBase$summarize(..., .dots = NULL)

Arguments

...

aggregation expressions in R syntax. Names of expressions are optional. If provided, names will be the fields of result arrayOp; otherwise field names are auto generated by scidb. Same syntax as ... in 'filter' and 'mutate' functions.

.dots

a list of aggregation expressions. Similar to '.dots' in 'mutate' and 'transmute'.

Returns

A new arrayOp instance

Method to_afl()

AFL string encapsulated by of the self ArrayOp

AFL can be either an scidb array name or array operation(s) on array(s).

The ArrayOp instance may have 'selected' fields but they are not reflected in the result. 'selected' fields are not reflected here, but determines which fields are retained in to_df() calls.

Usage

ArrayOpBase$to_afl()

Returns

an AFL expression string

Method to_schema_str()

Return a schema representation of the ArrayOp <attr1:type1 [, attr2:type2 ...]> [dim1 [;dim2]]

Unless sync_schema() is called, the schema may be inferred locally in R to save round trips between R and SciDB server. SciDB data frames have hidden dimensions that start with $

Usage

ArrayOpBase$to_schema_str()

Returns

SciDB schema string

Method to_df_all()

Download query result of self's AFL string with all self's fields.

Usage

ArrayOpBase$to_df_all()

Returns

An R data frame with columns from self's dimensions and attributes if no fields are selcted, or the selcted fields.

Method to_df()

Download query result of self's AFL string with all self's attributes.

Usage

ArrayOpBase$to_df()

Returns

An R data frame with columns from self's attributes if no fields are selected, or the selcted fields only.

Method execute()

Execute the AFL string for pure side effect without result returned.

Usage

ArrayOpBase$execute()

Returns

self

Method versions()

Return a data frame of all self's versions

Usage

ArrayOpBase$versions()

Returns

An R data frame with columns: version_id and timestamp

Method remove_versions()

Remove array versions of self

Only applicable to persistent arrays. Warning: This function will be executed effectively in scidb without extra 'execute()' and cannot be undone.

Usage

ArrayOpBase$remove_versions(version_id = NULL)

Arguments

version_id

NULL or a number. When set to NULL, all array versions are removed except for the latest one. When set to an number, must be a valid version_id of self, in which case all versions up to the 'version_id' are removed.

Returns

NULL

Method remove_array()

Remove array versions of self

Only applicable to persistent arrays. Warning: This function will be executed effectively in scidb without extra 'execute()' and cannot be undone.

Usage

ArrayOpBase$remove_array()

Returns

NULL

Method inner_join()

Inner join two arrays: 'self' (left) and 'right'

Similar to dplyr::inner_join, the result arrayOp performs an inner join. For both left and right arrays, only selected fields are included in the result arrayOp. If no fields are selected, then all fields are treated as selected.

Usage

ArrayOpBase$inner_join(
  right,
  on_left = NULL,
  on_right = NULL,
  on_both = NULL,
  left_alias = "_L",
  right_alias = "_R",
  join_mode = "equi_join",
  settings = NULL
)

Arguments

right

An arrayOp instance

on_left

NULL or a string vector as join keys. If set to NULL, join keys are inferred as shared fields of 'left' and 'right'.

on_right

NULL or a string vector as join keys. If set to NULL, join keys are inferred as shared fields of 'left' and 'right'.

on_both

NULL or a string vector as join keys. If set to NULL, join keys are inferred as shared fields of 'left' and 'right'. If not NULL, must be fields of both operands.

left_alias

Alias for left array to resolve potential conflicting fields in result

right_alias

Alias for right array to resolve potential conflicting fields in result

join_mode

String 'equi_join' or 'cross_join'. The latter requires join keys are all dimensions of both operands, which is more strigenet than 'equi_join' with the benifit of non-materializing result in scidb.

settings

A named list as join settings. E.g. list(algorithm = "'hash_replicate_right'")

Returns

A new arrayOp instance

Method left_join()

Left join two arrays: 'self' (left) and 'right'

Similar to dplyr::left_join, the result arrayOp performs a left join. For both left and right arrays, only selected fields are included in the result arrayOp. If no fields are selected, then all fields are treated as selected.

Usage

ArrayOpBase$left_join(
  right,
  on_left = NULL,
  on_right = NULL,
  on_both = NULL,
  left_alias = "_L",
  right_alias = "_R",
  settings = NULL
)

Arguments

right

An arrayOp instance

on_left

NULL or a string vector as join keys. If set to NULL, join keys are inferred as shared fields of 'left' and 'right'.

on_right

NULL or a string vector as join keys. If set to NULL, join keys are inferred as shared fields of 'left' and 'right'.

on_both

NULL or a string vector as join keys. If set to NULL, join keys are inferred as shared fields of 'left' and 'right'. If not NULL, must be fields of both operands.

left_alias

Alias for left array to resolve potential conflicting fields in result

right_alias

Alias for right array to resolve potential conflicting fields in result

settings

A named list as join settings. E.g. list(algorithm = "'hash_replicate_right'")

Returns

A new arrayOp instance

Method right_join()

Right join two arrays: 'self' (left) and 'right'

Similar to dplyr::right_join, the result arrayOp performs a right join. For both left and right arrays, only selected fields are included in the result arrayOp. If no fields are selected, then all fields are treated as selected.

Usage

ArrayOpBase$right_join(
  right,
  on_left = NULL,
  on_right = NULL,
  on_both = NULL,
  left_alias = "_L",
  right_alias = "_R",
  settings = NULL
)

Arguments

right

An arrayOp instance

on_left

NULL or a string vector as join keys. If set to NULL, join keys are inferred as shared fields of 'left' and 'right'.

on_right

NULL or a string vector as join keys. If set to NULL, join keys are inferred as shared fields of 'left' and 'right'.

on_both

NULL or a string vector as join keys. If set to NULL, join keys are inferred as shared fields of 'left' and 'right'. If not NULL, must be fields of both operands.

left_alias

Alias for left array to resolve potential conflicting fields in result

right_alias

Alias for right array to resolve potential conflicting fields in result

settings

A named list as join settings. E.g. list(algorithm = "'hash_replicate_right'")

Returns

A new arrayOp instance

Method semi_join()

Return an arrayOp instance with same schema as self and content cells that match the cells of 'df_or_arrayop'.

Similar to dplyr::semi_join, the result has the same schema as the left operand 'self' and with content filtered by 'df_or_arrayop'.

params field_mapping, lower_bound and upper_bound, if provided, must be named list, where the.names are from the source array (i.e. self), and values are from the right operand df_or_arrayop

Usage

ArrayOpBase$semi_join(
  df_or_arrayop,
  field_mapping = NULL,
  lower_bound = NULL,
  upper_bound = NULL,
  mode = "auto",
  filter_threshold = 200L,
  upload_threshold = 6000L
)

Arguments

df_or_arrayop

An R data frame or arrayOp instance.

mode

String of 'filter', 'cross_between', 'index_lookup' or 'auto'

filter_threshold

A number below which the 'filter' mode is used unless a mode other than 'auto' is provided.

upload_threshold

A number below which the 'df_or_arrayop' data frame is compield into a build literal array; otherwise uploaded to scidb as a regular array. Only applicable when 'df_or_arrayop' is an R data frame.

Returns

An arrayOp instance

Method persist()

Usage

ArrayOpBase$persist(save_array_name = NULL, .temp = FALSE, .gc = TRUE)

Method is_persistent()

Usage

ArrayOpBase$is_persistent()

Method exists_persistent_array()

Usage

ArrayOpBase$exists_persistent_array()

Method array_meta_data()

Usage

ArrayOpBase$array_meta_data()

Method finalize()

A finalize function executed when the 'self' instance is garbage collected in R

If an arrayOp is marked as .gc = T, then it should be removed from scidb when this function is executed.

Usage

ArrayOpBase$finalize()